+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-
+ - Added getting queues by priority range
+ 
 ## [3.8.1]
 ### Added
 - Added getting workers by range.